Move coordinator into tags
