+++ pyoozie/builder.py
@@ -2,7 +2,6 @@
 # Use of this source code is governed by a MIT-style license that can be found in the LICENSE file.
 from __future__ import unicode_literals
 
-from pyoozie import coordinator
 from pyoozie import tags
 
 
@@ -85,7 +84,7 @@ class CoordinatorBuilder(object):
     def __init__(self, name, workflow_xml_path, frequency_in_minutes, start, end=None, timezone=None,
                  workflow_configuration=None, timeout_in_minutes=None, concurrency=None, execution_order=None,
                  throttle=None, parameters=None):
-        self._coordinator = coordinator.Coordinator(
+        self._coordinator = tags.Coordinator(
             name=name,
             workflow_app_path=workflow_xml_path,
             frequency=frequency_in_minutes,
